06-mcp-http/fastmcp_examples/requirements.txt
# Requirements for FastMCP-based project
fastmcp==2.8.1
pyyaml==6.0.2
langgraph>=0.2.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.0
